REVIEW: fix review comments

+++ src/schedule/ucc_schedule_pipelined.c
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2021-2023, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * Copyright (c) 2021-2024, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
  *
  * See file LICENSE for terms.
  */
@@ -254,7 +254,7 @@ ucc_status_t ucc_dependency_handler(ucc_coll_task_t *parent,
                                     ucc_coll_task_t *task)
 {
     ucc_status_t status;
-    uint8_t      n_deps_satisfied;
+    uint32_t     n_deps_satisfied;
 
     n_deps_satisfied = ucc_atomic_fadd32(&task->n_deps_satisfied, 1);
     ucc_assert(task->n_deps_satisfied > n_deps_satisfied);
